requirements.txt
# Django dependencies
Django==6.0
djangorestframework==3.16.1
python-dotenv==1.2.1
requests==2.31.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
django-redis==5.4.0
django-cors-headers==4.3.1
whitenoise==6.6.0

# Bot dependencies
aiogram==3.13.1
aiohttp==3.10.5
aiofiles==24.1.0
asyncio-throttle==1.0.2

# System
psutil==6.0.0
colorlog==6.8.2
